[chore] 동행모집 화면전환 로직 추가

+++ feature/home/src/main/kotlin/com/tripmate/android/feature/home/navigation/HomeNavigation.kt
@@ -18,7 +18,7 @@ fun NavGraphBuilder.homeNavGraph(
     navigateToMateRecruit: () -> Unit,
     navigateToMateReview: () -> Unit,
     navigateToTripDetail: (spotId: String) -> Unit,
-    navigateToMateReviewPost: () -> Unit,
+    navigateToMateReviewPost: (Int) -> Unit,
     navigateToReport: () -> Unit,
 ) {
     composable(route = HOME_ROUTE) {
